refactor: HomeScreen
